+++ src/ts/renderer/utils/utils.types.ts
@@ -95,7 +95,7 @@ interface CenterModalData {
 interface CenterSearchModalData {
     readonly title: string;
     readonly onConfirm: (modal: CenterModal, searchResultContainerElement: HTMLElement) => Promise<ModalError>;
-    readonly onCancel?: () => void;
+    readonly onClose?: () => void;
     readonly onSearch: (searchResultContainerElement: HTMLElement, query: string) => Promise<void>;
     readonly searchDelay: number;
     readonly instantSearch: boolean;